requirements.txt
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
google-api-python-client